compat: don't fail if rdma/ib_verbs.h is missing
This is mainly a convenience patch for myself, still, this part is only
relevant when we (internally) build drbd_transport_rdma. In that case we
have the headers installed or it would fail badly anyways.

There are kernels/headers out there that don't come with rdma/ib_verbs.h
by default. For example Arch Linux is one of them, and I use Arch for
compat work because they are really fast (faster than Debian
sid/experimental) when it comes to packaging new kernel releases.