compat: introduce annotations
This adds annotations to compat tests. The annotation has to be:
- in the first line of test
- a one liner
- within a C comment
- valid json

It should at least have the "version" field set. For now, "author",
"commit", "comment", and "date" are valid too.

This makes it a lot easier to check which tests we can remove. We
already have a tool that prints all compat tests that have the same
value for all kernels we build for (i.e., candidates to remove). But
there are some that always have the same value because we do not
currently build for such new kernels. Obviously we can not remove these.
It is quite a PITA to check why and when a test was introduced, because
that information is scattered over git history and/or in file comments
(which vary). Having this machine readable allows output like this:

COMPAT_HAVE_BIOSET_INIT is always undefined (v:4.18, c:917a38c71af82185c39e31589587591fa764fb85)
COMPAT_HAVE_PROC_CREATE_SINGLE is always undefined (v:4.17, c:3f3942aca6da351a12543aa776467791b63b3a78)
COMPAT_HAVE_MAX_SEND_RECV_SGE is always undefined (v:4.18, c:33023fb85a42b53bf778bc025f9667b582282be4)

This makes it a lot easier to decide if this is a check worth looking
at.

Annotations got added for checks that findunused.py currently spits out.